Move formatting and coloring to formatting.rs
